backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.8.0
pydantic>=2.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
assemblyai>=0.30.0
